Use int64_t

+++ function.h
@@ -1,14 +1,14 @@
 #pragma once
 
-Function *get_func(Function **funcs, int funcs_size, char *func_name, int line) {
+Function *get_func(Function **funcs, int64_t funcs_size, char *func_name, int64_t line) {
     for (int i = 0; i < funcs_size; i++) {
         if (!strcmp(funcs[i]->name, func_name)) return funcs[i];
     }
-    ERR("ERROR on line %d: function %s not found\n", line, func_name);
+    ERR("ERROR on line %ld: function %s not found\n", line, func_name);
     return NULL;
 }
 
-int check_func(Function **funcs, int funcs_size, char *func_name) {
+int check_func(Function **funcs, int64_t funcs_size, char *func_name) {
     for (int i = 0; i < funcs_size; i++) {
         if (!strcmp(funcs[i]->name, func_name)) return i;
     }
